WIP: refactoring tests
