+++ typescript/optics-tests/test/home.test.ts
@@ -3,8 +3,9 @@ import { expect } from 'chai';
 import { getTestDeploy } from './testChain';
 import { OpticsState, Updater } from '../lib';
 import { Signer } from '../lib/types';
-import { Deploy } from '../../optics-deploy/src/chain';
-import * as deploys from '../../optics-deploy/src/deployOptics';
+import { CoreDeploy as Deploy } from '../../optics-deploy/src/deploy';
+import * as deploys from '../../optics-deploy/src/core';
+
 import {
   TestHome,
   UpdaterManager__factory,
